+++ ci/release_notes.md
@@ -1,1 +1,17 @@
+* Fixes to `eden catalog` table layout
+
+    ```console
+    Service                                              Plan                            Description
+    =======                                              ====                            ===========
+    searchify                                            small                           Small
+    ~                                                    plus                            Plus
+    ~                                                    pro                             Pro
+    blazemeter                                           free-tier                       Free
+    ~                                                    basic1kmr                       Basic
+    ~                                                    pro5kmr                         Pro
+    rediscloud                                           100mb                           Basic
+    ~                                                    250mb                           Mini
+    ~                                                    500mb                           Small
+    ```
+
 * Move to `dep` for dependencies